+++ inboxen/tests/settings.py
@@ -1,8 +1,13 @@
 from __future__ import absolute_import
 import os
 
+
+# tell settings module to ignore normal config file
 os.environ['INBOXEN_TESTING'] = '1'
 from inboxen.settings import *
+
+# build asserts if need be
+ASSETS_AUTO_BUILD = True
 
 CACHES = {
     "default": {
@@ -10,27 +15,17 @@ CACHES = {
     }
 }
 
-db = os.environ.get('DB', "sqlite")
-postgres_user = os.environ.get('PG_USER', 'postgres')
+# default to current user
+postgres_user = os.environ.get('PG_USER', os.environ.get('USER'))
 
 SECRET_KEY = "This is a test, you don't need secrets"
 ENABLE_REGISTRATION = True
 SECURE_SSL_REDIRECT = False
 
-if db == "sqlite":
-    DATABASES = {
-        'default': {
-            'ENGINE': 'django.db.backends.sqlite3',
-            'NAME': ':memory:',
-        },
-    }
-elif db == "postgres":
-    DATABASES = {
-        'default': {
-            'ENGINE': 'django.db.backends.postgresql_psycopg2',
-            'NAME': 'inboxen',
-            'USER': postgres_user,
-        },
-    }
-else:
-    raise NotImplementedError("Please check tests/settings.py for valid DB values")
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.postgresql_psycopg2',
+        'NAME': 'inboxenTest',
+        'USER': postgres_user,
+    },
+}